backend/requirements.txt
# Django & REST Framework
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0

# MongoDB
mongoengine>=0.28,<1.0

# File Parsing
PyMuPDF>=1.23,<2.0
python-docx>=1.1,<2.0

# OCR (for scanned/image-based PDFs)
pytesseract>=0.3.10,<1.0
Pillow>=10.0,<12.0

# NLP & AI
spacy>=3.7,<4.0
nltk>=3.8,<4.0
scikit-learn>=1.3,<2.0

# Utilities
python-dotenv>=1.0,<2.0
gunicorn>=21.2,<23.0

# Testing
pytest>=7.4,<9.0
